+++ package.json
@@ -32,7 +32,7 @@
     "g": "^2.0.1",
     "geist": "^1.3.0",
     "install": "^0.13.0",
-    "lucide-react": "^0.394.0",
+    "lucide-react": "^0.400.0",
     "next": "^14.2.4",
     "next-themes": "^0.3.0",
     "nextui-cli": "^0.3.2",
