+++ src/app/game/page.tsx
@@ -1,18 +1,12 @@
 "use client";
 
 import { GoogleAnalytics } from "@next/third-parties/google";
-import { ApiGoogle } from "../../components/apiGoogle/apiGoogle";
-import MenuTheme from "../../components/PageTheme";
-import Footer from "@/components/Footer";
-import Mode from "@/components/Footer/footerButton";
 import FooterButton from "@/components/Footer/footerButton";
-import TextLogo from "@/components/ContentPage/textLogo";
 import { TextInput } from "@/components/TextInput";
 import ProgressBar from "@/components/ProgressBar/index.";
 import Loading from "@/components/Loader";
 import React, { useState, useEffect } from "react";
-import { CorrectCard } from "@/components/correctCard";
-import { WrongCard } from "@/components/WrongCard";
+import { AudioIcon } from "@/components/Icons/audio";
 
 const PlayerPage = () => {
   const [isLoading, setIsLoading] = useState(true);
@@ -38,14 +32,28 @@ const PlayerPage = () => {
           progress={15}
           height="h-5"
           width="w-1/4"
-          className="mt-4 mx-auto" currentStep={0} totalSteps={5}        />
+          className="mt-4 mx-auto"
+          currentStep={0}
+          totalSteps={5}
+        />
       </div>
       <div className="pt-10 mt-10"></div>
-      <TextLogo />
+      <div className="flex flex-col items-center justify-center h-screen w-screen bg-white dark:bg-mirage-700">
+        <div className="text-center mb-auto mt-auto">
+          <div className="flex justify-center mb-4"></div>
+        </div>
+        <AudioIcon />
+        <h1 className="text-align: center font-bold text-4xl mt-2">
+          <span className="text-blue-400">
+            {" "}
+            Tap to <span className="text-black">hear </span>your phrase.
+          </span>
+        </h1>
+      </div>
       <div className="flex flex-col items-center p-5 mt-32">
         <TextInput onChange={undefined} />
       </div>
-   
+
       <FooterButton />
     </>
   );
