+++ src/service/phrases.tsx
@@ -1,0 +1,132 @@
+const phrases = {
+    backEnd: [
+      "Optimizing database queries for performance",
+      "Implementing RESTful APIs for seamless integration",
+      "Ensuring secure authentication and authorization",
+      "Refactoring legacy code for better maintainability",
+      "Scaling the server infrastructure for high traffic",
+      "Automating deployment with CI/CD pipelines",
+      "Monitoring server health with logging and alerts",
+      "Managing microservices with container orchestration",
+      "Integrating third-party services via APIs",
+      "Handling data migrations with minimal downtime",
+      "Optimizing database indexes for faster query performance",
+      "Implementing GraphQL endpoints for flexible data retrieval",
+      "Securing API endpoints against potential vulnerabilities",
+      "Refactoring monolithic applications into microservices",
+      "Designing caching strategies to improve response times",
+      "Automating database backups and disaster recovery",
+      "Analyzing server logs to optimize resource utilization",
+      "Deploying serverless functions for cost-efficient scaling",
+      "Integrating OAuth for third-party authentication",
+      "Managing secrets and environment variables securely",
+      "Designing robust database schemas for scalability",
+      "Implementing rate limiting to protect against abuse",
+      "Writing unit tests for server-side code",
+      "Managing load balancers for even traffic distribution",
+      "Configuring web servers for optimal performance",
+      "Implementing session management for user state persistence",
+      "Optimizing server response times with middleware",
+      "Handling concurrency in multi-threaded environments",
+      "Implementing data encryption for sensitive information",
+      "Utilizing NoSQL databases for flexible data storage",
+      "Managing distributed databases for global applications",
+      "Ensuring data integrity with ACID transactions",
+      "Implementing automated server provisioning with IaC tools",
+      "Building CI/CD pipelines for automated testing",
+      "Designing event-driven architectures for responsive systems",
+      "Utilizing message queues for asynchronous processing",
+      "Implementing full-text search capabilities",
+      "Optimizing server memory usage for efficiency",
+      "Handling API versioning for backward compatibility",
+      "Monitoring application performance with APM tools"
+    ],
+  
+    frontEnd: [
+      "Building responsive UIs with modern frameworks",
+      "Enhancing user experience with intuitive designs",
+      "Debugging front-end issues with browser dev tools",
+      "Implementing state management for complex apps",
+      "Optimizing load times with code splitting",
+      "Ensuring cross-browser compatibility",
+      "Using CSS preprocessors for better styling",
+      "Integrating front-end with back-end APIs",
+      "Creating interactive components with JavaScript",
+      "Testing UI components for consistency",
+      "Implementing lazy loading to improve initial page load times",
+      "Creating animations with CSS transitions and keyframes",
+      "Ensuring accessibility standards with ARIA roles",
+      "Using TypeScript for enhanced code integrity",
+      "Testing responsiveness across multiple device types",
+      "Optimizing SEO with structured data and meta tags",
+      "Integrating WebSockets for real-time updates",
+      "Utilizing WebAssembly to optimize performance-critical code",
+      "Implementing PWA features for offline functionality",
+      "Auditing performance metrics with Lighthouse",
+      "Creating pixel-perfect designs with CSS Grid and Flexbox",
+      "Utilizing component libraries for rapid UI development",
+      "Building reusable UI components for consistency",
+      "Implementing dark mode with CSS variables",
+      "Optimizing images for faster load times",
+      "Handling user input validation and feedback",
+      "Creating custom hooks in React for shared logic",
+      "Utilizing Vue.js for reactive data binding",
+      "Building forms with dynamic validation rules",
+      "Implementing internationalization (i18n) for global reach",
+      "Managing complex forms with Formik and Yup",
+      "Utilizing Webpack for asset bundling and optimization",
+      "Implementing drag-and-drop interfaces with JavaScript",
+      "Creating data visualizations with D3.js",
+      "Ensuring smooth scrolling with JavaScript",
+      "Building server-rendered applications with Next.js",
+      "Implementing authentication flows with JWT",
+      "Creating responsive typography with CSS units",
+      "Building custom themes with CSS-in-JS",
+      "Utilizing IndexedDB for client-side storage"
+    ],
+  
+    softSkills: [
+      "Communicating effectively with team members",
+      "Collaborating on code reviews for quality assurance",
+      "Managing time efficiently to meet deadlines",
+      "Adapting to new technologies and tools",
+      "Providing constructive feedback during meetings",
+      "Prioritizing tasks based on project goals",
+      "Mentoring junior developers to foster growth",
+      "Resolving conflicts with a positive attitude",
+      "Documenting code and processes for clarity",
+      "Balancing work and life for overall well-being",
+      "Facilitating productive team retrospectives for continuous improvement",
+      "Building consensus during decision-making processes",
+      "Adopting Agile methodologies to streamline project workflows",
+      "Demonstrating empathy in customer support interactions",
+      "Managing expectations with stakeholders effectively",
+      "Leading workshops to share best practices and insights",
+      "Promoting a culture of knowledge sharing within the team",
+      "Fostering diversity and inclusion initiatives",
+      "Staying informed about industry trends and best practices",
+      "Engaging in community outreach and mentorship programs",
+      "Leading team meetings with a clear agenda",
+      "Facilitating brainstorming sessions for creative solutions",
+      "Managing remote teams with effective communication tools",
+      "Encouraging peer-to-peer learning and collaboration",
+      "Handling feedback with an open mindset",
+      "Promoting team engagement through regular check-ins",
+      "Organizing team-building activities for morale boosting",
+      "Navigating cultural differences with sensitivity",
+      "Setting clear goals and expectations for team members",
+      "Fostering a growth mindset within the team",
+      "Encouraging innovation through a safe-to-fail environment",
+      "Providing timely recognition and appreciation",
+      "Conducting performance reviews with constructive feedback",
+      "Managing stress and preventing burnout",
+      "Encouraging work-life balance with flexible schedules",
+      "Building trust through transparency and integrity",
+      "Promoting a positive and inclusive team culture",
+      "Handling difficult conversations with tact and empathy",
+      "Aligning team efforts with organizational goals",
+      "Encouraging continuous learning and development"
+    ]
+  };
+
+
